Move Kubo-Greenwood to postprocess/localization
